+++ android/src/main/java/com/razorpay/razorpay_flutter/RazorpayDelegate.java
@@ -4,6 +4,9 @@ import android.app.Activity;
 import android.content.Intent;
 import android.util.Log;
 
+import androidx.annotation.NonNull;
+
+import com.google.gson.Gson;
 import com.razorpay.Checkout;
 import com.razorpay.CheckoutActivity;
 import com.razorpay.ExternalWalletListener;
@@ -13,7 +16,9 @@ import com.razorpay.PaymentResultWithDataListener;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import java.lang.reflect.Method;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.Map;
 
 import io.flutter.plugin.common.MethodChannel.Result;
@@ -38,14 +43,18 @@ public class RazorpayDelegate implements ActivityResultListener, ExternalWalletL
     private static final int INCOMPATIBLE_PLUGIN = 3;
     private static final int UNKNOWN_ERROR = 100;
     private String packageName;
+    private Checkout checkout;
+    Gson gson;
+    private UpiTurbo upiTurbo;
 
     public RazorpayDelegate(Activity activity) {
         this.activity = activity;
+        upiTurbo = new UpiTurbo(activity);
+        this.gson = new Gson();
     }
 
     void setPackageName(String packageName) {
         this.packageName = packageName;
-        Log.d("PackageName", packageName);
     }
 
     void openCheckout(Map<String, Object> arguments, Result result) {
@@ -54,7 +63,6 @@ public class RazorpayDelegate implements ActivityResultListener, ExternalWalletL
 
         JSONObject options = new JSONObject(arguments);
         if (activity.getPackageName().equalsIgnoreCase(packageName)) {
-            Log.d("PAYMENT", activity.getPackageName() + ";;;" + packageName);
             Intent intent = new Intent(activity, CheckoutActivity.class);
             intent.putExtra("OPTIONS", options.toString());
             intent.putExtra("FRAMEWORK", "flutter");
@@ -102,10 +110,34 @@ public class RazorpayDelegate implements ActivityResultListener, ExternalWalletL
 
         Map<String, Object> data = new HashMap<>();
         data.put("code", translateRzpPaymentError(code));
-        data.put("message", message);
+        try {
+            JSONObject response = new JSONObject(message);
+            JSONObject errorObj = response.getJSONObject("error");
+            data.put("message", errorObj.getString("description"));
+            JSONObject metadata = errorObj.getJSONObject("metadata");
+            Map<String, String> metadataHash = new HashMap<>();
+            Iterator<String> metaKeys = metadata.keys();
+            while (metaKeys.hasNext()) {
+                String key = metaKeys.next();
+                metadataHash.put(key, metadata.getString(key));
+            }
+            errorObj.remove("metadata");
+            Map<String, Object> resp = new HashMap<>();
+            Iterator<String> keys = errorObj.keys();
+            while (keys.hasNext()) {
+                String key = keys.next();
+                resp.put(key, errorObj.get(key));
+            }
+            resp.put("metadata", metadataHash);
+            resp.put("email", paymentData.getUserEmail());
+            resp.put("contact", paymentData.getUserContact());
+            data.put("responseBody", resp);
+        } catch (JSONException e) {
+            data.put("message", message);
+            data.put("responseBody", message);
+        }
 
         reply.put("data", data);
-
         sendReply(reply);
     }
 
@@ -123,17 +155,22 @@ public class RazorpayDelegate implements ActivityResultListener, ExternalWalletL
             try {
                 data.put("razorpay_subscription_id", paymentData.getData().optString("razorpay_subscription_id"));
             } catch (Exception e) {
-                e.printStackTrace();
             }
         }
-
         reply.put("data", data);
         sendReply(reply);
     }
 
     @Override
     public boolean onActivityResult(int requestCode, int resultCode, Intent data) {
-        Checkout.handleActivityResult(activity, requestCode, resultCode, data, this, this);
+        try {
+            Method merchantActivityResult = Checkout.class.getMethod("merchantActivityResult", Activity.class,
+                    Integer.class, Integer.class, Intent.class, PaymentResultWithDataListener.class,
+                    ExternalWalletListener.class);
+            merchantActivityResult.invoke(null, activity, requestCode, resultCode, data, this, this);
+        } catch (Exception e) {
+            Checkout.handleActivityResult(activity, requestCode, resultCode, data, this, this);
+        }
         return true;
     }
 
@@ -147,6 +184,22 @@ public class RazorpayDelegate implements ActivityResultListener, ExternalWalletL
         reply.put("data", data);
 
         sendReply(reply);
+    }
+
+    public void setKeyID(String keyId, Result result) {
+        upiTurbo.setKeyID(keyId, result);
+    }
+
+    public void linkNewUpiAccount(String customerMobile, String color, Result result) {
+        upiTurbo.linkNewUpiAccount(customerMobile, color, result);
+    }
+
+    public void manageUpiAccounts(String customerMobile, String color, Result result) {
+        upiTurbo.manageUpiAccounts(customerMobile, color, result);
+    }
+
+    public boolean isTurboPluginAvailable(Result result) {
+        return upiTurbo.isTurboPluginAvailable(result);
     }
 
 }